% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aircraftOPS-package.R
\docType{package}
\name{aircraftOPS}
\alias{aircraftOPS}
\alias{aircraftOPS-package}
\title{aircraftOPS: R code for use with aircraft operation data}
\description{
The aircraftOPS package was created as part of PhD 
work by the lead author. Package code was developed for use 
with AIMS and aircraft operations data.
}
\section{aircraftOPS functions}{

Packaged Functions include: readAIMSHTMLAttachment, extractHTMLAttachmentFromMSG.
}

