% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractHTMLAttachmentFromMSG.R
\name{extractHTMLAttachmentFromMSG}
\alias{extractHTMLAttachmentFromMSG}
\title{extractHTMLAttachmentFromMSG}
\usage{
extractHTMLAttachmentFromMSG(file = file.choose(), output = NULL)
}
\arguments{
\item{file}{The file to be read. By default, this opens a file selection browser.}

\item{output}{Output require, by default.}
}
\value{
The function returns an AIMS SMS attachment.
}
\description{
extractHTMLAttachmentFromMSG extracts attachments from AIMS SMS messages.
}
\examples{
#to do
}
\author{
Karl Ropkins (2015/07/09, v 0.3)
}
\keyword{methods}
