% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAIMSHTMLAttachment.R
\name{readAIMSHTMLAttachment}
\alias{readAIMSHTMLAttachment}
\title{readAIMSHTMLAttachment}
\usage{
readAIMSHTMLAttachment(file = file.choose(), make.numeric = c("fob.on",
  "fob.off"), verbose = FALSE)
}
\arguments{
\item{file}{The file to be read. By default, this opens a file selection browser.}

\item{make.numeric}{The names of any data columns which, if found, should be converted to numerics.}

\item{verbose}{If TRUE, the function provides a full read report.}
}
\value{
A data.frame of AIMS flight reports.
}
\description{
readAIMSHTMLAttachment reads data in AIMS SMS message attachments.
}
\examples{
#to do
}
\author{
Karl Ropkins (2015/07/09, v 0.3)
}
\keyword{methods}
